fix : Console과 inputValidator 분리 후 Calculator 안으로 캡슐화
