+++ src/main/kotlin/calculator/CalculatorApplication.kt
@@ -1,20 +1,24 @@
 package calculator
 
 import calculator.domain.Calculator
+import calculator.io.Console
 
 fun main() {
     Main().start()
 }
 
 class Main(
-    private val calculator: Calculator,
+    private val console: Console,
+    private val calculator: Calculator
 ) {
 
     constructor() : this(
+        Console(),
         Calculator()
     )
 
     fun start() {
-        calculator.run()
+        val input = console.getInput()
+        calculator.run(input)
     }
 }
