+++ src/main/kotlin/calculator/domain/Calculator.kt
@@ -1,19 +1,17 @@
 package calculator.domain
 
-import calculator.io.Console
 import calculator.utils.CalculatorParser
 
 class Calculator {
 
-    private val console: Console = Console()
+    private val inputValidator: InputValidator = InputValidator()
 
-    fun run() {
-        val input = console.getInput()
-        val extractedList = CalculatorParser.extract(input)
+    fun run(input: String?) {
+        val validatedInput = inputValidator.validate(input)
+        val extractedList = CalculatorParser.extract(validatedInput)
         val result = calculate(extractedList)
         println(result)
     }
-
 
     /**
      * 추출한 문자열 계산
@@ -23,7 +21,7 @@ class Calculator {
         var result = list[0].toDouble()
 
         for (i in 1 until list.size step 2) {
-            val operator = Operator.findOperate(list[i][0])
+            val operator = Operator.findOperate(list[i])
             result = operator.operate(result, list[i + 1].toDouble())
         }
