+++ src/main/kotlin/calculator/io/Console.kt
@@ -1,5 +1,7 @@
 package calculator.io
 
+import calculator.domain.InputValidator
+
 /**
  * 사용자 입력 받기
  */
