+++ src/test/kotlin/calculator/domain/InputValidatorTest.kt
@@ -1,4 +1,4 @@
-package calculator.io
+package calculator.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
